fix negative filesizes on 32-bit systems

+++ src/Filesystem.php
@@ -326,7 +326,7 @@ class Filesystem implements FilesystemInterface
             return false;
         }
 
-        return (int) $object['size'];
+        return (float) $object['size'];
     }
 
     /**
